# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Slogr Controller API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/analytics/{sid}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session details",
                "description": "Retrieve session details by ID",
                "operationId": "c3d27cc1ec71a5940edaf64a9ab24cef",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "analytic": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "client": {
                                                    "type": "object"
                                                },
                                                "server": {
                                                    "type": "object"
                                                },
                                                "rtt": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "up": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "down": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "profile": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agentlogs/{id}": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Get agent logs",
                "description": "Retrieve agent logs",
                "operationId": "b3b7adb58fb6e3fb0f50cdb62bfb40d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "logs": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "300": {
                        "description": "Client Unreachable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agentdata/{id}/{profile}": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Get agent data",
                "description": "Retrieve agent data",
                "operationId": "d293cee9c8945df838ec78895128b3ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pro",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "res": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "agent": {
                                                    "type": "object"
                                                },
                                                "pros": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "profile": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "300": {
                        "description": "Client Unreachable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "security": {
                "passport": {
                    "type": "oauth2",
                    "description": "Laravel passport oauth2 security.",
                    "in": "header",
                    "scheme": "https",
                    "flows": {
                        "password": {
                            "authorizationUrl": "localhost/slogr/oauth/authorize",
                            "tokenUrl": "localhost/slogr/oauth/token",
                            "refreshUrl": "localhost/slogr/token/refresh",
                            "scopes": []
                        }
                    }
                }
            }
        }
    }
}